feat: use docker for pigeon and file generation
This is the first attempt at using docker for generating files so that
we only need docker for generating files and we can build and pin the
tools used to generate files without using dep.

This is useful because it means we don't have to add the dependency to
dep and it works for non-go dependencies. This will make it trivial to
transition to using go modules so we can drop dep as a dependency.